requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
google-generativeai
pandas
matplotlib
python-dotenv
Pillow
werkzeug
requests
psycopg2-binary
reportlab
